requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
sqlparse==0.5.3
tzdata==2025.2
asgiref==3.8.1
black==23.12.1
boto3==1.37.33
botocore==1.37.33
Brotli==1.1.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
distlib==0.3.7
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
firebase-admin==6.6.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
msgpack==1.1.0
mypy-extensions==1.0.0
numpy==1.26.3
packaging==24.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==3.11.0
proto-plus==1.25.0
protobuf==5.28.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyotp==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
s3transfer==0.11.4
setuptools==69.5.1
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.24.6
virtualenvwrapper-win==1.2.7
wheel==0.43.0
whitenoise==6.8.2
asgiref==3.8.1
black==23.12.1
boto3==1.37.33
botocore==1.37.33
Brotli==1.1.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
distlib==0.3.7
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
firebase-admin==6.6.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jmespath==1.0.1
msgpack==1.1.0
mypy-extensions==1.0.0
numpy==1.26.3
packaging==24.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==3.11.0
proto-plus==1.25.0
protobuf==5.28.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyotp==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
s3transfer==0.11.4
setuptools==69.5.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.24.6
virtualenvwrapper-win==1.2.7
wheel==0.43.0
whitenoise==6.8.2
